openni2: Move method definitions out of headers
