+++ src/openni2/ColorStream.hpp
@@ -46,7 +46,7 @@ namespace Freenect2Driver
 
   private:
     typedef std::map< OniVideoMode, int > FreenectVideoModeMap;
-    OniSensorType getSensorType() const { return ONI_SENSOR_COLOR; }
+    OniSensorType getSensorType() const;
     VideoModeMap getSupportedVideoModes() const;
     void populateFrame(libfreenect2::Frame* srcFrame, int srcX, int srcY, OniFrame* dstFrame, int dstX, int dstY, int width, int height) const;
     
@@ -59,91 +59,11 @@ namespace Freenect2Driver
     ColorStream(libfreenect2::Freenect2Device* pDevice, Freenect2Driver::Registration *reg);
     //~ColorStream() { }
 
-    OniStatus setImageRegistrationMode(OniImageRegistrationMode mode)
-    {
-      if (mode == ONI_IMAGE_REGISTRATION_DEPTH_TO_COLOR) {
-        // XXX, switch color resolution to 512x424 for registrarion here
-        OniVideoMode video_mode = makeOniVideoMode(ONI_PIXEL_FORMAT_RGB888, 512, 424, 30);
-        setProperty(ONI_STREAM_PROPERTY_VIDEO_MODE, &video_mode, sizeof(video_mode));
-      }
-      return ONI_STATUS_OK;
-    }
+    OniStatus setImageRegistrationMode(OniImageRegistrationMode mode);
 
     // from StreamBase
-    OniBool isPropertySupported(int propertyId)
-    {
-      switch(propertyId)
-      {
-        default:
-          return VideoStream::isPropertySupported(propertyId);
-          
-        case ONI_STREAM_PROPERTY_HORIZONTAL_FOV:
-        case ONI_STREAM_PROPERTY_VERTICAL_FOV:
-        case ONI_STREAM_PROPERTY_AUTO_WHITE_BALANCE:
-        case ONI_STREAM_PROPERTY_AUTO_EXPOSURE:
-          return true;
-      }
-    }
-
-    OniStatus getProperty(int propertyId, void* data, int* pDataSize)
-    {
-      switch (propertyId)
-      {
-        default:
-          return VideoStream::getProperty(propertyId, data, pDataSize);
-
-        case ONI_STREAM_PROPERTY_HORIZONTAL_FOV:     // float (radians)
-        {
-          if (*pDataSize != sizeof(float))
-          {
-            LogError("Unexpected size for ONI_STREAM_PROPERTY_HORIZONTAL_FOV");
-            return ONI_STATUS_ERROR;
-          }
-          *(static_cast<float*>(data)) = HORIZONTAL_FOV;
-          return ONI_STATUS_OK;
-        }
-        case ONI_STREAM_PROPERTY_VERTICAL_FOV:       // float (radians)
-        {
-          if (*pDataSize != sizeof(float))
-          {
-            LogError("Unexpected size for ONI_STREAM_PROPERTY_VERTICAL_FOV");
-            return ONI_STATUS_ERROR;
-          }
-          *(static_cast<float*>(data)) = VERTICAL_FOV;
-          return ONI_STATUS_OK;
-        }
-        
-        // camera
-        case ONI_STREAM_PROPERTY_AUTO_WHITE_BALANCE: // OniBool
-        {
-          if (*pDataSize != sizeof(OniBool))
-          {
-            LogError("Unexpected size for ONI_STREAM_PROPERTY_AUTO_WHITE_BALANCE");
-            return ONI_STATUS_ERROR;
-          }
-          *(static_cast<OniBool*>(data)) = auto_white_balance;
-          return ONI_STATUS_OK;
-        }
-        case ONI_STREAM_PROPERTY_AUTO_EXPOSURE:      // OniBool
-        {
-          if (*pDataSize != sizeof(OniBool))
-          {
-            LogError("Unexpected size for ONI_STREAM_PROPERTY_AUTO_EXPOSURE");
-            return ONI_STATUS_ERROR;
-          }
-          *(static_cast<OniBool*>(data)) = auto_exposure;
-          return ONI_STATUS_OK;
-        }
-      }
-    }
-    
-    OniStatus setProperty(int propertyId, const void* data, int dataSize)
-    {
-      switch (propertyId)
-      {
-        default:
-          return VideoStream::setProperty(propertyId, data, dataSize);
-      }
-    }
+    OniBool isPropertySupported(int propertyId);
+    OniStatus getProperty(int propertyId, void* data, int* pDataSize);
+    OniStatus setProperty(int propertyId, const void* data, int dataSize);
   };
 }
