+++ src/openni2/Registration.hpp
@@ -24,6 +24,8 @@
  * either License.
  */
 
+#pragma once
+
 #include <libfreenect2/registration.h>
 
 namespace Freenect2Driver {
@@ -40,7 +42,7 @@ namespace Freenect2Driver {
 
     void depthFrame(libfreenect2::Frame* frame);
     void colorFrameRGB888(libfreenect2::Frame* srcFrame, libfreenect2::Frame* dstFrame);
-    void setEnable(bool enable = true) { enabled = enable; }
-    bool isEnabled() { return enabled; }
+    void setEnable(bool enable = true);
+    bool isEnabled();
   };
 }
